rename applications.utils to applications.core
