+++ cstar/applications/roms_marbl/transforms.py
@@ -15,8 +15,8 @@ from pydantic import (
     model_validator,
 )
 
+from cstar.applications.core import Transform
 from cstar.applications.roms_marbl.models import RomsMarblBlueprint
-from cstar.applications.utils import Transform
 from cstar.base.exceptions import CstarExpectationFailed
 from cstar.base.utils import DEFAULT_OUTPUT_ROOT_NAME, deep_merge, slugify
 from cstar.orchestration.orchestration import LiveStep
