+++ cstar/tests/conftest.py
@@ -4,8 +4,8 @@ from unittest.mock import Mock, patch
 
 import pytest
 
+from cstar.applications.core import register_application
 from cstar.applications.roms_marbl.app import RomsMarblApplication
-from cstar.applications.utils import register_application
 
 
 @register_application
